fastapi/requirements.txt
fastapi
uvicorn
openai-whisper
ffmpeg-python
python-multipart
pydub
whisper
numpy
aiofiles
python-dotenv